+++ sdk/communication/Azure.Communication.CallAutomation/src/CallAutomationClientOptions.cs
@@ -33,7 +33,6 @@ namespace Azure.Communication.CallAutomation
             ApiVersion = version switch
             {
                 ServiceVersion.V2023_03_06 => "2023-03-06",
-                ServiceVersion.V2023_06_15_Preview => "2023-06-15-preview",
                 ServiceVersion.V2023_10_15 => "2023-10-15",
                 _ => throw new ArgumentOutOfRangeException(nameof(version)),
             };
@@ -51,14 +50,9 @@ namespace Azure.Communication.CallAutomation
             V2023_03_06 = 1,
 
             /// <summary>
-            /// The beta2 of the CallAutomation service.
-            /// </summary>
-            V2023_06_15_Preview = 2,
-
-            /// <summary>
             /// The GA2 of the CallAutomation service.
             /// </summary>
-            V2023_10_15 = 3
+            V2023_10_15 = 2
 #pragma warning restore CA1707 // Identifiers should not contain underscores
         }
     }