+++ sdk/communication/Azure.Communication.CallAutomation/src/Models/Events/MediaEventReasonCode.cs
@@ -35,6 +35,7 @@ namespace Azure.Communication.CallAutomation
 
         private const string PlayDownloadFailedValue = "8536";
         private const string PlayInvalidFileFormatValue = "8535";
+        private const string PlaySourceTextOrSsmlEmptyValue = "8582";
         private const string CognitiveServicesErrorValue = "8565";
 
         private const string CompletedSuccessfullyValue = "0";
@@ -68,8 +69,10 @@ namespace Azure.Communication.CallAutomation
 
         /// <summary> Action failed, file could not be downloaded. </summary>
         public static MediaEventReasonCode PlayDownloadFailed { get; } = new MediaEventReasonCode(PlayDownloadFailedValue);
-        /// <summary> Action failed, file could not be downloaded. </summary>
+        /// <summary> Action failed, invalid file format. </summary>
         public static MediaEventReasonCode PlayInvalidFileFormat { get; } = new MediaEventReasonCode(PlayInvalidFileFormatValue);
+        /// <summary> Action failed, play source text or ssml empty. </summary>
+        public static MediaEventReasonCode PlaySourceTextOrSsmlEmpty { get; } = new MediaEventReasonCode(PlaySourceTextOrSsmlEmptyValue);
         /// <summary> Action failed, cognitive service error. </summary>
         public static MediaEventReasonCode CognitiveServicesError { get; } = new MediaEventReasonCode(CognitiveServicesErrorValue);