+++ sdk/communication/Azure.Communication.CallAutomation/src/Models/SendDtmfTonesOptions.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -36,8 +37,9 @@ namespace Azure.Communication.CallAutomation
         public string OperationContext { get; set; }
 
         /// <summary>
-        /// The callback URI to override the main callback URI.
+        /// The callback URI that overrides the default callback URI set by CreateCall/AnswerCall for this operation.
+        /// This setup is per-action. If this is not set, the default callback URI set by CreateCall/AnswerCall will be used.
         /// </summary>
-        public string OverrideCallbackUrl { get; set; }
+        public Uri OperationCallbackUri { get; set; }
     }
 }